requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
filelock==3.14.0
fsspec==2024.5.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.0
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
pillow==10.3.0
proglog==0.1.10
pytube==15.0.0
regex==2024.5.15
requests==2.32.3
sympy==1.12.1
tbb==2021.12.0
tiktoken==0.7.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
